feat: accept promise in next.config
NextJS already allows developers to specify a function in `next.config.js` to
configure their application. Instead of throwing an error, this change
allows users of the framework to write a function that returns a
promise. This is useful for situations where you might want to run some
async code in order to figure out how to configure the application at
build time, like fetching secret configuration or a myriad of other
tasks.

Closes #24837